fix(health): cap NB at 3 neighbors, hex-encode DUID keys, fix uint8_t overflow
- getNeighborCacheJson: cap neighbor output at 3 entries (was 8)
- getNeighborCacheJson: hex-encode DUID key to prevent raw bytes from
  corrupting the pipe-delimited format or surrounding JSON
- prepareForSending: change data_length from uint8_t to size_t to prevent
  silent wrap-around for payloads >= 256 bytes bypassing the size check

+++ src/CdpPacket.cpp
@@ -1,7 +1,7 @@
 #include "CdpPacket.h"
 
 int CdpPacket::prepareForSending() {
-    uint8_t data_length = data.size();
+    size_t data_length = data.size();
     buffer.clear();
 
     if ( data.empty() || data_length > MAX_DATA_LENGTH) {
